requirements.txt
streamlit
pymupdf
streamlit
pandas
matplotlib
PyMuPDF
pdfplumber
google-generativeai
protobuf

